# BackendServer/demo/api.yaml
openapi: 3.0.0
info:
    title: FDS
    description: |
        This API allows to communicate with FDS (Food Delivery System) server.
    version: 0.0.8
paths:
...
    /vendor/register:
        post:
            summary: Register a new vendor
            description: Create a new vendor account with unique username, strong password for authentication and other vendor info.
            operationId: registerVendor
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/vendorDetails'
                required: true
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sessionId'
                "409":
                    description: "Conflict"
                    content:
                        application/json:
                            schema:
                                type: string
                                enum:
                                    - Username
                                    - Phone
                                    - Email
                                    - Password
    /vendor/login:
        post:
            summary: Login to vendor account
            description: Get username and password and authenticate the vendor.
                Returns sessionId for further requests
            operationId: loginVendor
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - username
                                - password
                            properties:
                                username:
                                    type: string
                                password:
                                    type: string
                required: true
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sessionId'
                "403":
                    description: "Forbidden"
    /vendor/product/add:
        post:
            summary:
            description:
            operationId: vendorAddProduct
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - name
                                - price
                            properties:
                                name:
                                    type: string
                                thumbnail:
                                    type: string
                                price:
                                    type: number
                                maxQuantity:
                                    type: integer
                                imageUrls:
                                    type: array
                                    items:
                                        type: string
                                        minLength: 40
                                        maxLength: 40
                                tags:
                                    type: array
                                    description: tags related to food item
                                    items:
                                        type: string
                required: true
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/itemId'
                "403":
                    description: "Forbidden"
    /vendor/product/edit:
        post:
            summary:
            description:
            operationId: vendorEditProduct
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - itemId
                            properties:
                                itemId:
                                    $ref: '#/components/schemas/itemId'
                                name:
                                    type: string
                                thumbnail:
                                    type: string
                                price:
                                    type: number
                                maxQuantity:
                                    type: integer
                required: true
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/foodItemFull'
                "403":
                    description: "Forbidden"
    /vendor/product/changeAvailable:
        post:
            summary:
            description:
            operationId: vendorChangeProductAvailabile
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - itemId
                                - status
                            properties:
                                itemId:
                                    $ref: '#/components/schemas/itemId'
                                status:
                                    description: Weather the item is available or not
                                    type: boolean
                required: true
            responses:
                "200":
                    description: "OK"
                "403":
                    description: "Forbidden"
    /vendor/changeAvailable:
        post:
            summary:
            description:
            operationId: vendorChangeAvailabile
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - itemId
                                - status
                            properties:
                                itemId:
                                    $ref: '#/components/schemas/itemId'
                                status:
                                    description: Weather the item is available or not
                                    type: boolean
                required: true
            responses:
                "200":
                    description: "OK"
                "403":
                    description: "Forbidden"
    /vendor/order/requested:
        get:
            summary:
            description:
            operationId: vendorGetRequestedOrders
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        orderId:
                                            type: string
                                        userId:
                                            type: string
                                        vendorID:
                                            type: string
                                        pickupLocation:
                                            type: string
                                        orderItems:
                                            type: array
                                            items:
                                                type: object
                                                required:
                                                    - itemId
                                                    - quantity
                                                properties:
                                                    itemId:
                                                        type: string
                                                        minLength: 40
                                                        maxLength: 40
                                                    quantity:
                                                        type: integer
                                                        description: Quantity of the specific item.
                                        location:
                                            type: string
                                            description: Delivery location.
                                        price:
                                            type: number
                "403":
                    description: "Forbidden"
    /vendor/order/accepted:
        get:
            summary:
            description:
            operationId: vendorGetAcceptedOrders
            parameters:
            - name: sessionId
              in : header
              required: true
              schema:
                  $ref: '#/components/schemas/sessionId'
            responses:
                "200":
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        orderId:
                                            type: string
                                        userId:
                                            type: string
                                        vendorID:
                                            type: string
                                        pickupLocation:
                                            type: string
                                        orderItems:
                                            type: array
                                            items:
                                                type: object
                                                required:
                                                    - itemId
                                                    - quantity
                                                properties:
                                                    itemId:
                                                        type: string
                                                        minLength: 40
                                                        maxLength: 40
                                                    quantity:
                                                        type: integer
                                                        description: Quantity of the specific item.
                                        location:
                                            type: string
                                            description: Delivery location.
                                        price:
                                            type: number
                "403":
                    description: "Forbidden"
...
components:
    schemas:
        sessionId:
            type: string
            maxLength: 40
            minLength: 40
        itemId:
            type: string
            maxLength: 40
            minLength: 40
...
        foodItemFull:
            type: object
            properties:
                itemId:
                    type: string
                    description: ID of the item
                itemName:
                    type: string
                    description: Name of the item
                thumbnailPicture:
                    type: string
                    description: 64bit encoded thumbnail picture
                price:
                    type: number
                    description: Price of the item
                vendorName:
                    type: string
                    description: Name of the vendor
                vendorLocation:
                    type: string
                    description: Location of the vendor
                starRating:
                    type: number
                    description: Star rating of the item
                isAvailable:
                    type: boolean
                    description: Indicates if the item is available
                maxQuantity:
                    type: integer
                    description: Maximum quantity available
                imageUrls:
                    type: array
                    description: URLs of additional images
                    items:
                        type: string
                tags:
                    type: array
                    description: tags related to food item
                    items:
                        type: string
...
        vendorDetails:
            type: object
            required:
                - username
                - password
                - phone
                - location
            properties:
                username:
                    type: string
                    example: "testVendor1"
                password:
                    type: string
                    example: "1234@Ved"
                name:
                    type: string
                    example: "UltraTech Cafe"
                phone:
                    type: string
                    example: "9878543210"
                email:
                    type: string
                    example: "vendor001@example.com"
                location:
                    type: string
                    example: "Mess Block"
...
servers:
    - url: http://13.233.99.87:8080